+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 
+## 0.0.4
+
+* 新增-打开设置 应用详情页api，支持指定包名 - Added, open app setting detail, support assign app package name.
+* 扩展-打开应用市场应用api支持指定应用 - Expand, open appmarket api, support assign app package name.
+* 新增-内置kotlin plugin, 无需外部添加 - Added, built in kotlin plugin.
+
 ## 0.0.3
+
 * downloadAndInstallApk & installApk 新增弹窗提示文案字段 - Added popup prompt text field.
 * downloadAndInstallApk 新增下载状态变化时回调 - Added callback when download status changes.